Slit app/start.go into options and config
